\name{fr3PG-package}
\alias{fr3PG-package}
\alias{fr3PG}
\docType{package}
\title{
\packageTitle{fr3PG}
}
\description{
\packageDescription{fr3PG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fr3PG}
\packageIndices{fr3PG}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fr3PG}

Maintainer: \packageMaintainer{fr3PG}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
